// apps/api/src/modules/official-documents/pdf.types.ts
export interface CertificateData {
  institutionName: string;
  documentTitle: string;
  studentName: string;
  documentNumber: string;
  studentCode: string;
  program: string;
  semester: number;
  trackingNumber: string;
  issuedAt: string;
  version: number;
}

export interface ConstancyData {
  institutionName: string;
  documentTitle: string;
  studentName: string;
  documentNumber: string;
  studentCode: string;
  program: string;
  semester: number;
  purpose: string;
  trackingNumber: string;
  issuedAt: string;
  version: number;
}